moved podium to its own component
