+++ client/src/views/Host.vue
@@ -59,6 +59,18 @@ export default {
       // playerlist contains player objects for every connected player
       playerList: [
         {
+            name: 'Safi',
+            color: 'orange',
+            pfp: '8',
+            clientId: 4
+          },
+           {
+            name: 'Yona',
+            color: 'yellow',
+            pfp: '7',
+            clientId: 5
+          },
+        {
           name: 'Jack',
           color: 'red',
           pfp: '2',
@@ -81,7 +93,9 @@ export default {
           color: 'yellow',
           pfp: '7',
           clientId: 6
-        }  
+        },
+       
+          
       ],
       // prompt responses each round are stored here. response obj format { player: player, response: string }
       promptResponses: [],
@@ -91,42 +105,6 @@ export default {
       totalRounds: 3,
       // responses that won in voting round
       winningResponses: [
-        {
-          player: {
-            name: 'Jack',
-            color: 'red',
-            pfp: '2',
-            clientId: 1
-          },
-          response: 'I am a response'
-        },
-        {
-          player: {
-            name: 'Josh',
-            color: 'blue',
-            pfp: '3',
-            clientId: 2
-          },
-          response: 'I am a response'
-        },
-        {
-          player: {
-            name: 'Thomas',
-            color: 'green',
-            pfp: '6',
-            clientId: 3
-          },
-          response: 'I am a response'
-        },
-        {
-          player: {
-            name: 'Chris',
-            color: 'yellow',
-            pfp: '7',
-            clientId: 6
-          },
-          response: 'I am a response'
-        }
       ],
       // true if page is not visible (using visibilitychange event listener)
       isPageHidden: false,
